requirements.txt
sounddevice
soundfile
openai-whisper
google-generativeai
streamlit
pyrubberband 
datasets 
sentence-transformers 
faiss-cpu